Corrected non-critical comments
